assert test helpers for testing mixins
`assert_style`, `assert_partial` and `assert_template`.  Use these
to verify your mixin applies its stuff as you expect.

Note that this does limited testing as there is no convenient way
to test the content of each thing, just the args used to create
each thing.  We'll need to add a spy for each type of element/style
to be able to test content.